backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0.38,<3.0
alembic>=1.14,<2.0
pydantic-settings>=2.8,<3.0
jinja2>=3.1,<4.0
boto3>=1.37,<2.0
botocore>=1.37,<2.0
playwright>=1.52,<2.0
httpx>=0.28,<1.0
